Move logging.* tests to io.quarkus.logging package
