+++ core/test-extension/deployment/src/test/java/io/quarkus/logging/PeriodicSizeRotatingLoggingTest.java
@@ -1,4 +1,4 @@
-package logging;
+package io.quarkus.logging;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,7 +11,7 @@ import java.util.logging.Logger;
 
 import org.jboss.logmanager.formatters.PatternFormatter;
 import org.jboss.logmanager.handlers.DelayedHandler;
-import org.jboss.logmanager.handlers.PeriodicRotatingFileHandler;
+import org.jboss.logmanager.handlers.PeriodicSizeRotatingFileHandler;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.junit.jupiter.api.Test;
@@ -20,15 +20,15 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import io.quarkus.runtime.logging.InitialConfigurator;
 import io.quarkus.test.QuarkusUnitTest;
 
-public class PeriodicRotatingLoggingTest {
+public class PeriodicSizeRotatingLoggingTest {
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest()
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
-                    .addAsResource("application-periodic-file-log-rotating.properties", "application.properties"))
-            .setLogFileName("PeriodicRotatingLoggingTest.log");
+                    .addAsResource("application-periodic-size-file-log-rotating.properties", "application.properties"))
+            .setLogFileName("PeriodicSizeRotatingLoggingTest.log");
 
     @Test
-    public void periodicRotatingConfigurationTest() {
+    public void periodicSizeRotatingConfigurationTest() {
         LogManager logManager = LogManager.getLogManager();
         assertThat(logManager).isInstanceOf(org.jboss.logmanager.LogManager.class);
 
@@ -36,7 +36,7 @@ public class PeriodicRotatingLoggingTest {
         assertThat(Logger.getLogger("").getHandlers()).contains(delayedHandler);
 
         Handler handler = Arrays.asList(delayedHandler.getHandlers()).stream()
-                .filter(h -> (h instanceof PeriodicRotatingFileHandler))
+                .filter(h -> (h instanceof PeriodicSizeRotatingFileHandler))
                 .findFirst().get();
         assertThat(handler).isNotNull();
         assertThat(handler.getLevel()).isEqualTo(Level.INFO);
@@ -45,5 +45,8 @@ public class PeriodicRotatingLoggingTest {
         assertThat(formatter).isInstanceOf(PatternFormatter.class);
         PatternFormatter patternFormatter = (PatternFormatter) formatter;
         assertThat(patternFormatter.getPattern()).isEqualTo("%d{HH:mm:ss} %-5p [%c{2.}]] (%t) %s%e%n");
+
+        PeriodicSizeRotatingFileHandler periodicSizeRotatingFileHandler = (PeriodicSizeRotatingFileHandler) handler;
+        assertThat(periodicSizeRotatingFileHandler.isRotateOnBoot()).isFalse();
     }
 }
